Added docker-compose

+++ project/app/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Client(models.Model):
 
     username = models.CharField(max_length=255, blank=True, null=True)
-    spent_money = models.PositiveSmallIntegerField(blank=True, null=True)
+    spent_money = models.PositiveBigIntegerField(blank=True, null=True)
     gems = models.JSONField(blank=True, null=True)
 
     def __str__(self):
